app/pyproject.toml
[project]
name = "thaleia"
version = "0.1.0"
description = "Beautiful reports, born from your data."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2",
    "django-ninja>=1.4.0",
    "kaleido==0.2.0",
    "plotly>=6.0.1",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.0",
    "weasyprint>=65.0",
]

[dependency-groups]
dev = [
 "ruff>=0.11.4",
]
